feat: extract console into separate component
