+++ web/src/components/preview/Console/index.ts
@@ -1,0 +1,1 @@
+export * from './Console'
